% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretRFRepeatedCV.R
\name{caretRFRepeatedCV}
\alias{caretRFRepeatedCV}
\title{Random Forest with Repeated Cross Validation}
\arguments{
\item{x}{\code{Phyloseq} object with \code{sample_data} and \code{otu_table}}

\item{features}{Features to use for RF model.}

\item{sample_group}{Variable in \code{sample_data} to
test for differences in taxa abundances}

\item{folds}{Repeated hold-out to apply. Default is '5'}

\item{repeats}{Number of repeats. Default is '100'}

\item{seed}{Variable in \code{sample_data} to
test for differences in taxa abundances}

\item{metric}{See \code{\link[caret]{train}}. Default is 'Accuracy'}

\item{print_final}{Logical. Default is TRUE. Prints the final model output}

\item{...}{Additional arguments passed on to \code{\link[caret]{train}}.}
}
\value{
A list is returned of class train. See \code{\link[caret]{train}}
}
\description{
Random Forest with Repeated Cross Validation
}
\details{
Performs a Random forest with repeated Cross-validation. The function
         as the name suggests uses caret under the hood. Make sure to cite the
         original \code{caret} package with correct version number.
}
\examples{

library(biomeStats)
library(microbiome)
library(biomeUtils)
ps <- FuentesIliGutData |>
             microbiome::transform("compositional") |>
             biomeUtils::mutateTaxaTable(FeatureID = taxa_names(FuentesIliGutData)) |>
             biomeUtils::filterSampleData(ILI != "L2")
# select features reduced for speed in example
features.to.use <- core_members(ps, 0.01, 25/100)

# for example reduce folds and repeats
rf.fit <- caretRFRepeatedCV(ps,
                            sample_group = "ILI",
                            features = features.to.use,
                            folds = 3,
                            repeats = 5,
                            set.seed = 1819,
                            metric = "Accuracy",
                            print_final=TRUE)
print(rf.fit)
#Check splits for each fold
#table(rf.fit$pred$Resample, rf.fit$pred$obs)

}
\seealso{
\code{\link[caret]{train}}
}
\author{
Sudarshan A. Shetty
}
