% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check.strata}
\alias{check.strata}
\title{check.strata}
\usage{
check.strata(s, data.to.test)
}
\arguments{
\item{s}{Stratum that is created to check for in the analysis}

\item{data.to.test}{dataset variable to specifically test}
}
\description{
check.strata
}
