% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_association.R
\name{check_association}
\alias{check_association}
\title{Check for taxa-metadata associations}
\usage{
check_association(
  data.for.testing,
  B = 1e+05,
  no.of.factors = NULL,
  aux.columns = NULL,
  log.scale = TRUE,
  nominal.bound.on.FDR = 0.25,
  phen.data = NULL,
  compare.label = "case_control",
  aux.TYPES = NULL,
  cat.types = NULL,
  path_loc = ".",
  name.of.stratification = "Sex and age-group",
  verbose = TRUE
)
}
\arguments{
\item{data.for.testing}{A data frame combined from transform_to_fractions and prep_metadata.}

\item{B}{Resampling default to 100000}

\item{no.of.factors}{Number of factor from `data.for.testing` to test.}

\item{aux.columns}{specific column numbers with taxa abundances.}

\item{log.scale}{Logical. Plot scale. Default and suggested TRUE}

\item{nominal.bound.on.FDR}{cut-off for multiple testing. Default=0.25}

\item{phen.data}{data frame output from prep_metadata()}

\item{compare.label}{label to save files.}

\item{aux.TYPES}{How to treat each of the variables in the data.for.testing}

\item{cat.types}{Categorical to plot proportions positive}

\item{path_loc}{Location to store/save output}

\item{name.of.stratification}{label of stratum e.g."Sex and age-group"}

\item{verbose}{TRUE}
}
\description{
Check for taxa-metadata associations
}
\details{
The associations between taxa and metadata variables are tested. These are done using
the method/appraoched described by JA Ferreira and S Fuentes https://doi.org/10.1093/bib/bbz077
}
