% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create.names}
\alias{create.names}
\title{Create variable names}
\usage{
create.names(data.set)
}
\arguments{
\item{data.set}{dataset consisting of variables}
}
\description{
Create variable names
}
