% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAssociationAnalysis.R
\name{doAssociationAnalysis}
\alias{doAssociationAnalysis}
\title{Check for taxa-metadata associations}
\usage{
doAssociationAnalysis(
  data_for_testing,
  B = 1e+05,
  no_of_factors = NULL,
  aux_columns = NULL,
  log_scale = TRUE,
  nominal_bound_on_FDR = 0.25,
  phen_data = NULL,
  compare_label = "case_control",
  aux_TYPES = NULL,
  cat_ypes = NULL,
  path_loc = ".",
  name_of_stratification = "Sex and age-group",
  verbose = TRUE
)
}
\arguments{
\item{data_for_testing}{A data frame combined from prepTaxaAbundanceData and prepSampleData}

\item{B}{Resampling default to 100000}

\item{no_of_factors}{Number of factor from `data_for_testing` to test.}

\item{aux_columns}{specific column numbers with taxa abundances.}

\item{log_scale}{Logical. Plot scale. Default and suggested TRUE}

\item{nominal_bound_on_FDR}{cut-off for multiple testing. Default=0.25}

\item{phen_data}{data frame output from prepSampleData()}

\item{compare_label}{label to save files.}

\item{aux_TYPES}{How to treat each of the variables in the data_for_testing}

\item{cat_ypes}{Categorical to plot proportions positive}

\item{path_loc}{Location to store/save output}

\item{name_of_stratification}{label of stratum e.g."Sex and age-group"}

\item{verbose}{TRUE}
}
\description{
Check for taxa-metadata associations
}
\details{
The associations between taxa and metadata variables are tested.
         These are done using the method/approach described by
         JA Ferreira and S Fuentes
         \url{https://doi.org/10.1093/bib/bbz077}
}
\references{
Ferreira JA, Fuentes S. (2020). Some comments on certain statistical aspects of
the study of the microbiome.
\emph{Briefings in bioinformatics} 21(4), pp.1487-1494.
\url{https://doi.org/10.1093/bib/bbz077}
}
\author{
Sudarshan A. Shetty
}
