% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doKruskalWallisTest.R
\name{doKruskalWallisTest}
\alias{doKruskalWallisTest}
\title{Do Kruskal-Wallis Test Taxa and Multiple Groups}
\usage{
doKruskalWallisTest(
  x,
  sample_group = NULL,
  adj_method = "BH",
  effect_size = TRUE,
  ...
)
}
\arguments{
\item{x}{A phyloseq object}

\item{sample_group}{Variable in \code{sample_data} to
test for differences in taxa abundances}

\item{adj_method}{One of "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none". Default="BH".
For more information see
\code{\link[rstatix]{adjust_pvalue}}}

\item{effect_size}{Logical. Default is TRUE. Calculates effect size for
Kruskal-Wallis test using
\code{\link[rstatix]{kruskal_effsize}}}

\item{...}{Additional parameters that are passed on to
\code{\link[rstatix]{kruskal_effsize}}}
}
\value{
A tibble.
}
\description{
Do Kruskal-Wallis Test Taxa and Multiple Groups
}
\details{
Performs Kruskal-Wallis Test one or multiple levels and
         is a wrapper using \code{\link[rstatix]{kruskal_test}},
         \code{\link[rstatix]{adjust_pvalue}} and
         \code{\link[rstatix]{add_significance}} for an
         input \code{phyloseq} object. Useful for comparing
         large number of taxa between groups.
}
\examples{
library(biomeStats)
library(phyloseq)
library(microbiome)
library(dplyr)
data("FuentesIliGutData")
ps.rel <- microbiome::transform(FuentesIliGutData, "compositional")
ps.rel <- core(ps.rel, detection=0.001, prevalence=50/100)

kw_results <- doKruskalWallisTest(ps.rel,
                                  sample_group = "ILI",
                                  adj_method = "BH",
                                  effect_size = TRUE)

print(kw_results)

}
\seealso{
\code{\link[rstatix]{kruskal_test}}
         \code{\link[rstatix]{kruskal_effsize}}
}
\author{
Sudarshan A. Shetty
}
