% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBlockedWilcoxTest.R, R/doWilcoxTest.R
\name{doWilcoxTest}
\alias{doWilcoxTest}
\alias{doBlockedWilcoxTest}
\title{Do a Blocked Wilcox Test Taxa on Groups}
\usage{
doBlockedWilcoxTest(
  x,
  features = NULL,
  sample_group = NULL,
  block_fct = NULL,
  adj_method = "BH",
  ...
)

doWilcoxTest(x, sample_group = NULL, paired = FALSE, adj_method = "BH", ...)
}
\arguments{
\item{x}{A phyloseq object}

\item{features}{Features to test}

\item{sample_group}{Variable in \code{sample_data} to
test for differences in taxa abundances}

\item{block_fct}{Column variable to stratify/block in testing.}

\item{adj_method}{One of "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none". Default="BH".
For more information see
\code{\link[rstatix]{adjust_pvalue}}.}

\item{...}{Additional parameters that are passed on to
\code{\link[rstatix]{wilcox_test}}}

\item{paired}{Logical TRUE or FALSE passed to
\code{\link[rstatix]{wilcox_test}}}
}
\value{
A tibble.

A tibble.
}
\description{
Do a Blocked Wilcox Test Taxa on Groups

Do Wilcox Test Taxa and Paire-wise on Groups
}
\details{
Performs one and two sample blocked Wilcoxon tests
         wrapper using \code{\link[coin]{wilcox_test}}.
         Useful for comparing large number of taxa between groups while
         stratifying by confounder like sex, age, etc.

Performs one and two sample Wilcoxon tests
         wrapper using \code{\link[rstatix]{wilcox_test}},
         \code{\link[rstatix]{adjust_pvalue}} and
         \code{\link[rstatix]{add_significance}} for an
         input \code{phyloseq} object. Useful for comparing
         large number of taxa between groups.
}
\examples{
library(biomeStats)
library(biomeUtils)
library(phyloseq)
library(microbiome)
library(dplyr)
data("FuentesIliGutData")
feat.to.test <- core_members(getProportions(FuentesIliGutData), 0.01, 25/100)
ps <- getProportions(FuentesIliGutData) |>
  filterSampleData(ILI !="L2") |>
  filterTaxaByNames(ids = feat.to.test, keep = TRUE)

wilcox_results <- doBlockedWilcoxTest(x = ps,
                                      features = feat.to.test,
                                      sample_group = "ILI",
                                      block_fct = "sex",
                                      adj_method = "BH")
print(wilcox_results)

library(biomeStats)
library(phyloseq)
library(microbiome)
library(dplyr)
data("FuentesIliGutData")
ps <- subset_samples(FuentesIliGutData, ILI \%in\% c("C", "L1"))
ps.rel <- microbiome::transform(ps, "compositional")
ps.rel <- core(ps.rel, detection=0.001, prevalence=50/100)

wilcox_results <- doWilcoxTest(ps.rel,
                               sample_group = "ILI",
                               paired = FALSE,
                               adj_method = "BH",
                               alternative = "greater") \%>\%
                  dplyr::filter(p.adj <= 0.05)
print(wilcox_results)

}
\seealso{
\code{\link[rstatix]{wilcox_test}}

\code{\link[rstatix]{wilcox_test}}
}
\author{
Sudarshan A. Shetty
}
