% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ili_data}
\alias{ili_data}
\title{Test genus abundances and metadata}
\format{
Phyloseq object
\itemize{
\item phyloseq data
\code{\link{phyloseq-class}} object.
}
}
\usage{
data(ili_data)
}
\value{
Loads the data set in R.
}
\description{
Genus level counts and example metadata.
For details, see \url{XXXX}.
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
