% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrate.ONE.association.BY.STRATUM.R
\name{illustrate.ONE.association.BY.STRATUM}
\alias{illustrate.ONE.association.BY.STRATUM}
\title{Illustrate.ONE.association.BY.STRATUM}
\usage{
illustrate.ONE.association.BY.STRATUM(
  data.for.plotting,
  TYPES,
  stratum,
  x.limits,
  y.limits,
  log.scale
)
}
\arguments{
\item{data.for.plotting}{dataset to use for plotting}

\item{TYPES}{variable properties, ordinal, categorical, continous, etc.}

\item{stratum}{stratum to check}

\item{x.limits}{x-axis limits for plotting}

\item{y.limits}{y-axis limits for plotting}

\item{log.scale}{use log-scale to plot abundances}
}
\description{
Illustrate.ONE.association.BY.STRATUM
}
\details{
Plots association blocked by stratum
}
