% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrate.ONE.association.R
\name{illustrate.ONE.association}
\alias{illustrate.ONE.association}
\title{Illustrate.ONE.association}
\usage{
illustrate.ONE.association(columns, data.for.plotting, types, log.scale)
}
\arguments{
\item{columns}{dataset variable to specifically test}

\item{data.for.plotting}{dataset to use for plotting}

\item{types}{variable properties, ordinal, categorical, continous, etc.}

\item{log.scale}{use log-scale to plot abundances}
}
\description{
Illustrate.ONE.association
}
\details{
Plots association
}
