% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{show.missing.by.variable}
\alias{show.missing.by.variable}
\title{show.missing.by.variable}
\usage{
show.missing.by.variable(data.set, plot)
}
\arguments{
\item{data.set}{dataset consisting of variables}

\item{plot}{Plot}
}
\description{
show.missing.by.variable
}
