% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.ONE.association.0.R
\name{test.ONE.association.0}
\alias{test.ONE.association.0}
\title{Illustrate.ONE.association.BY.STRATUM}
\usage{
test.ONE.association.0(columns, data.for.testing, TYPES, variables.of.interest)
}
\arguments{
\item{columns}{dataset variable to specifically test}

\item{data.for.testing}{dataset to use for testing}

\item{TYPES}{variable properties, ordinal, categorical, continous, etc.}

\item{variables.of.interest}{variables to test}
}
\description{
Illustrate.ONE.association.BY.STRATUM
}
\details{
Tests based on the whole data set blocked by stratum.
This version uses the asymptotic distribution
}
