% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.ONE.association.R
\name{test.ONE.association}
\alias{test.ONE.association}
\title{Test one association}
\usage{
test.ONE.association(
  columns,
  data.for.testing,
  TYPES,
  variables.of.interest,
  B.i
)
}
\arguments{
\item{columns}{dataset variable to specifically test}

\item{data.for.testing}{dataset to use for testing}

\item{TYPES}{variable properties, ordinal, categorical, continous, etc.}

\item{variables.of.interest}{list/vector of variables matching input colnames to test}

\item{B.i}{resample values (numeric)}
}
\description{
Test one association
}
\details{
Tests based on the whole data set blocked by stratum
}
\references{
Ferreira JA, Fuentes S. (2020). Some comments on certain statistical aspects of
the study of the microbiome.
\emph{Briefings in bioinformatics} 21(4), pp.1487-1494.
\url{https://doi.org/10.1093/bib/bbz077}
}
\author{
Sudarshan A. Shetty
}
