% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_fractions.R
\name{transform_to_fractions}
\alias{transform_to_fractions}
\title{Transform to Fractions}
\usage{
transform_to_fractions(x, det.thres = 0.001, prev.thres = 5/100)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{det.thres}{Remove taxa with less than this abundance in total dataset}

\item{prev.thres}{Remove taxa with less than this prevalence in total dataset}
}
\value{
A tibble
}
\description{
Transform to Fractions
}
\details{
Prepares the \code{\link{phyloseq-class}} object to
         taxa abundance table for testing.
}
\examples{
library(biomestats)
data("ili_data")
ps <- ili_data
gen_tab <- transform_to_fractions(ps,
                                  det.thres= 0.0001,
                                  prev.thres=5/100)

}
\references{
Ferreira JA, Fuentes S. (2020). Some comments on certain statistical aspects of
the study of the microbiome.
\emph{Briefings in bioinformatics} 21(4), pp.1487-1494.
\url{https://doi.org/10.1093/bib/bbz077}
}
\author{
Sudarshan A. Shetty
}
