% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Wilson.interval}
\alias{Wilson.interval}
\title{Wilson.interval}
\usage{
Wilson.interval(frequency, n, confidence)
}
\arguments{
\item{frequency}{frequency}

\item{n}{Number must be numeric}

\item{confidence}{confidence value}
}
\description{
Wilson.interval
}
