% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doWilcoxTest.R
\name{doWilcoxTest}
\alias{doWilcoxTest}
\title{Do Wilcox Test Taxa and Paire-wise on Groups}
\usage{
doWilcoxTest(x, sample_group = NULL, paired = FALSE, adj_method = "BH", ...)
}
\arguments{
\item{x}{A phyloseq object}

\item{sample_group}{Variable in \code{sample_data} to
test for differences in taxa abundances}

\item{paired}{Logical TRUE or FALSE passed to
\code{\link[rstatix]{wilcox_test}}}

\item{adj_method}{One of "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none". Default="BH".
For more information see
\code{\link[rstatix]{adjust_pvalue}}.}

\item{...}{Additional parameters that are passed on to
\code{\link[rstatix]{wilcox_test}}}
}
\value{
A tibble.
}
\description{
Do Wilcox Test Taxa and Paire-wise on Groups
}
\details{
Performs one and two sample Wilcoxon tests
         wrapper using \code{\link[rstatix]{wilcox_test}},
         \code{\link[rstatix]{adjust_pvalue}} and
         \code{\link[rstatix]{add_significance}} for an
         input \code{phyloseq} object. Useful for comparing
         large number of taxa between groups.
}
\examples{
library(biomeStats)
library(phyloseq)
library(microbiome)
library(dplyr)
data("FuentesIliGutData")
ps <- subset_samples(FuentesIliGutData, ILI \%in\% c("C", "L1"))
ps.rel <- microbiome::transform(ps, "compositional")
ps.rel <- core(ps.rel, detection=0.0001, prevalence=50/100)

wilcox_results <- doWilcoxTest(ps.rel,
                               sample_group = "ILI",
                               paired = FALSE,
                               adj_method = "BH",
                               alternative = "greater") \%>\%
                  dplyr::filter(p.adj <= 0.05)
print(wilcox_results)

}
\seealso{
\code{\link[rstatix]{wilcox_test}}
}
\author{
Sudarshan A. Shetty
}
