% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrate.ONE.association.BY.STRATUM.R
\name{illustrate.ONE.association.BY.STRATUM}
\alias{illustrate.ONE.association.BY.STRATUM}
\title{Illustrate.ONE.association.BY.STRATUM}
\usage{
illustrate.ONE.association.BY.STRATUM(
  data.for.plotting,
  TYPES,
  stratum,
  x.limits,
  y.limits,
  log_scale
)
}
\arguments{
\item{data.for.plotting}{dataset to use for plotting}

\item{TYPES}{variable properties, ordinal, categorical, continous, etc.}

\item{stratum}{stratum to check}

\item{x.limits}{x-axis limits for plotting}

\item{y.limits}{y-axis limits for plotting}

\item{log_scale}{use log-scale to plot abundances}
}
\description{
Illustrate.ONE.association.BY.STRATUM
}
\details{
Plots association blocked by stratum
}
\references{
Ferreira JA, Fuentes S. (2020). Some comments on certain statistical aspects of
the study of the microbiome.
\emph{Briefings in bioinformatics} 21(4), pp.1487-1494.
\url{https://doi.org/10.1093/bib/bbz077}
}
\author{
Sudarshan A. Shetty
}
