% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrate.ONE.association.R
\name{illustrate.ONE.association}
\alias{illustrate.ONE.association}
\title{Illustrate.ONE.association}
\usage{
illustrate.ONE.association(columns, data.for.plotting, types, log_scale)
}
\arguments{
\item{columns}{dataset variable to specifically test}

\item{data.for.plotting}{dataset to use for plotting}

\item{types}{variable properties, ordinal, categorical, continous, etc.}

\item{log_scale}{use log-scale to plot abundances}
}
\description{
Illustrate.ONE.association
}
\details{
Plots association
}
\references{
Ferreira JA, Fuentes S. (2020). Some comments on certain statistical aspects of
the study of the microbiome.
\emph{Briefings in bioinformatics} 21(4), pp.1487-1494.
\url{https://doi.org/10.1093/bib/bbz077}
}
\author{
Sudarshan A. Shetty
}
