% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.ONE.association.0.R
\name{test.ONE.association.0}
\alias{test.ONE.association.0}
\title{Illustrate.ONE.association.BY.STRATUM}
\usage{
test.ONE.association.0(columns, data_for_testing, TYPES, variables.of.interest)
}
\arguments{
\item{columns}{dataset variable to specifically test}

\item{data_for_testing}{dataset to use for testing}

\item{TYPES}{variable properties, ordinal, categorical, continuous, etc.}

\item{variables.of.interest}{variables to test}
}
\description{
Illustrate.ONE.association.BY.STRATUM
}
\details{
Tests based on the whole data set blocked by stratum.
         This version uses the asymptotic distribution
}
\references{
Ferreira JA, Fuentes S. (2020). Some comments on certain statistical aspects of
the study of the microbiome.
\emph{Briefings in bioinformatics} 21(4), pp.1487-1494.
\url{https://doi.org/10.1093/bib/bbz077}
}
\author{
Sudarshan A. Shetty
}
