% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_fractions.R
\name{transform_to_fractions}
\alias{transform_to_fractions}
\title{Transform to Fractions}
\usage{
transform_to_fractions(x, filter.abund = 0.1)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{filter.abund}{remove taxa with less than this abundance in total dataset}
}
\description{
Transform to Fractions
}
\details{
Prepares the taxa abundance table for testing.
}
